Separated object insertions
