+++ Import_Yml.py
@@ -1,7 +1,7 @@
 
 import FreeCAD as App , Mesh , Part
 
-from Accessors import getTransparency , getPlacement , getRotation , getColor
+from Source.Objects import *
 from builtins import open as openFile
 from yaml import safe_load
 from os import path
@@ -19,212 +19,6 @@ def insertObject(directory, filename, document, group, attributes = None):
     if filename[-4:] in ['.stp', '.igs', 'iges', 'step']:
         return insertPart(directory, filename, document, group, attributes)
     insertMesh(directory, filename, document, group, attributes)
-
-def insertMesh(directory, filename, document, group, attributes = None):
-    mesh = Mesh.Mesh(path.join(directory,filename))
-    object_name = filename[:-4]
-    if 'objectName' in attributes:
-        object_name = attributes['objectName']
-    new_mesh = document.addObject("Mesh::Feature", object_name)
-    new_mesh.Mesh = mesh
-    if attributes:
-        color = getColor(attributes)
-        if color:
-            new_mesh.ViewObject.ShapeColor = color
-        transparency = getTransparency(attributes)
-        if transparency:
-            new_mesh.ViewObject.Transparency = transparency
-        placement = getPlacement(attributes)
-        rotation = getRotation(attributes)
-        new_mesh.Placement = App.Placement(placement, rotation)
-    group.addObject(new_mesh)
-
-def insertPart(directory, filename, document, group, attributes = None):
-    if not path.isfile(path.join(directory, filename)):
-        directory = path.expanduser('~/.FreeCAD/Mod/yaml-workspace')
-        if not path.isfile(path.join(directory, filename)):
-            print(f'ERROR: `{ filename }` not found!')
-            return
-
-    part = Part.Shape()
-    part = Part.read(path.join(directory,filename))
-    object_name = filename[:-4]
-    if 'objectName' in attributes:
-        object_name = attributes['objectName']
-    new_part = document.addObject("Part::Feature", object_name)
-    new_part.Shape = part
-    if attributes:
-        color = getColor(attributes)
-        if color:
-            new_part.ViewObject.ShapeColor = color
-        transparency = getTransparency(attributes)
-        if transparency:
-            new_part.ViewObject.Transparency = transparency
-        placement = getPlacement(attributes)
-        rotation = getRotation(attributes)
-        new_part.Placement = App.Placement(placement, rotation)
-    group.addObject(new_part)
-
-def insertCylinder(name, document, group, attributes):
-    solid = document.addObject("Part::Cylinder","Cylinder")
-    solid.Label = name
-    solid.Radius = f'{ attributes[ "radius" ] } mm'
-    solid.Height = f'{ attributes[ "height" ] } mm'
-    if 'angle' in attributes:
-        solid.Angle = f'{ attributes[ "angle" ] } deg'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertSphere(name, document, group, attributes):
-    solid = document.addObject("Part::Sphere","Sphere")
-    solid.Label = name
-    solid.Radius = f'{ attributes[ "radius" ] } mm'
-    if 'angle1' in attributes:
-        solid.Angle1 = f'{ attributes[ "angle1" ] } deg'
-    if 'angle2' in attributes:
-        solid.Angle2 = f'{ attributes[ "angle2" ] } deg'
-    if 'angle3' in attributes:
-        solid.Angle3 = f'{ attributes[ "angle3" ] } deg'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertEllipsoid(name, document, group, attributes):
-    solid = document.addObject("Part::Ellipsoid","Ellipsoid")
-    solid.Label = name
-    solid.Radius1 = f'{ attributes[ "radius1" ] } mm'
-    solid.Radius2 = f'{ attributes[ "radius2" ] } mm'
-    solid.Radius3 = f'{ attributes[ "radius3" ] } mm'
-    if 'angle1' in attributes:
-        solid.Angle1 = f'{ attributes[ "angle1" ] } deg'
-    if 'angle2' in attributes:
-        solid.Angle2 = f'{ attributes[ "angle2" ] } deg'
-    if 'angle3' in attributes:
-        solid.Angle3 = f'{ attributes[ "angle3" ] } deg'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertBox(name, document, group, attributes):
-    solid = document.addObject("Part::Box","Box")
-    solid.Label = name
-    solid.Length = f'{ attributes[ "length" ] } mm'
-    solid.Height = f'{ attributes[ "height" ] } mm'
-    solid.Width = f'{ attributes[ "width" ] } mm'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertCone(name, document, group, attributes):
-    solid = document.addObject("Part::Cone","Cone")
-    solid.Label = name
-    solid.Radius1 = f'{ attributes[ "radius1" ] } mm'
-    solid.Radius2 = f'{ attributes[ "radius2" ] } mm'
-    solid.Height = f'{ attributes[ "height" ] } mm'
-    if 'angle' in attributes:
-        solid.Angle = f'{ attributes[ "angle" ] } deg'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertTorus(name, document, group, attributes):
-    solid = document.addObject("Part::Torus","Torus")
-    solid.Label = name
-    solid.Radius1 = f'{ attributes[ "radius1" ] } mm'
-    solid.Radius2 = f'{ attributes[ "radius2" ] } mm'
-    if 'angle1' in attributes:
-        solid.Angle1 = f'{ attributes[ "angle1" ] } deg'
-    if 'angle2' in attributes:
-        solid.Angle2 = f'{ attributes[ "angle2" ] } deg'
-    if 'angle3' in attributes:
-        solid.Angle3 = f'{ attributes[ "angle3" ] } deg'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertPrism(name, document, group, attributes):
-    solid = document.addObject("Part::Prism","Prism")
-    solid.Label = name
-    solid.Polygon = int(attributes[ "polygon" ])
-    solid.Circumradius = f'{ attributes[ "radius" ] } mm'
-    solid.Height = f'{ attributes[ "height" ] } mm'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
-
-def insertWedge(name, document, group, attributes):
-    solid = document.addObject("Part::Wedge","Wedge")
-    solid.Label = name
-    solid.Xmin = f'{ attributes[ "xmin" ] } mm'
-    solid.Ymin = f'{ attributes[ "ymin" ] } mm'
-    solid.Zmin = f'{ attributes[ "zmin" ] } mm'
-    solid.X2min = f'{ attributes[ "x2min" ] } mm'
-    solid.Z2min = f'{ attributes[ "z2min" ] } mm'
-    solid.Xmax = f'{ attributes[ "xmax" ] } mm'
-    solid.Ymax = f'{ attributes[ "ymax" ] } mm'
-    solid.Zmax = f'{ attributes[ "zmax" ] } mm'
-    solid.X2max = f'{ attributes[ "x2max" ] } mm'
-    solid.Z2max = f'{ attributes[ "z2max" ] } mm'
-    color = getColor(attributes)
-    if color:
-        solid.ViewObject.ShapeColor = color
-    transparency = getTransparency(attributes)
-    if transparency:
-        solid.ViewObject.Transparency = transparency
-    placement = getPlacement(attributes)
-    rotation = getRotation(attributes)
-    solid.Placement = App.Placement(placement, rotation)
-    group.addObject(solid)
