+++ Source/Objects/Box.py
@@ -1,0 +1,22 @@
+
+import FreeCAD as App
+
+from Source.Accessors import *
+
+
+def insertBox(name, document, group, attributes):
+    solid = document.addObject("Part::Box","Box")
+    solid.Label = name
+    solid.Length = f'{ attributes[ "length" ] } mm'
+    solid.Height = f'{ attributes[ "height" ] } mm'
+    solid.Width = f'{ attributes[ "width" ] } mm'
+    color = getColor(attributes)
+    if color:
+        solid.ViewObject.ShapeColor = color
+    transparency = getTransparency(attributes)
+    if transparency:
+        solid.ViewObject.Transparency = transparency
+    placement = getPlacement(attributes)
+    rotation = getRotation(attributes)
+    solid.Placement = App.Placement(placement, rotation)
+    group.addObject(solid)
