+++ Source/Objects/Sphere.py
@@ -1,0 +1,26 @@
+
+import FreeCAD as App
+
+from Source.Accessors import *
+
+
+def insertSphere(name, document, group, attributes):
+    solid = document.addObject("Part::Sphere","Sphere")
+    solid.Label = name
+    solid.Radius = f'{ attributes[ "radius" ] } mm'
+    if 'angle1' in attributes:
+        solid.Angle1 = f'{ attributes[ "angle1" ] } deg'
+    if 'angle2' in attributes:
+        solid.Angle2 = f'{ attributes[ "angle2" ] } deg'
+    if 'angle3' in attributes:
+        solid.Angle3 = f'{ attributes[ "angle3" ] } deg'
+    color = getColor(attributes)
+    if color:
+        solid.ViewObject.ShapeColor = color
+    transparency = getTransparency(attributes)
+    if transparency:
+        solid.ViewObject.Transparency = transparency
+    placement = getPlacement(attributes)
+    rotation = getRotation(attributes)
+    solid.Placement = App.Placement(placement, rotation)
+    group.addObject(solid)
