+++ Source/Objects/__init__.py
@@ -1,0 +1,11 @@
+
+from Source.Objects.Ellipsoid import *
+from Source.Objects.Cylinder import *
+from Source.Objects.Sphere import *
+from Source.Objects.Prism import *
+from Source.Objects.Torus import *
+from Source.Objects.Wedge import *
+from Source.Objects.Cone import *
+from Source.Objects.Mesh import *
+from Source.Objects.Part import *
+from Source.Objects.Box import *
